Update handle data sidebar
